Fix
Renaming of class date fields has been performed.
Changed property types for date classes.

+++ app/src/main/java/ru/netology/fmhandroid/dao/PatientDao.kt
@@ -13,7 +13,7 @@ interface PatientDao {
     fun getAllPatients(): Flow<List<PatientEntity>>
 
     @Query("SELECT * FROM PatientEntity WHERE id = :id")
-    suspend fun getPatientById(id: Long): PatientEntity
+    suspend fun getPatientById(id: Int): PatientEntity
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(patient: PatientEntity)
